Search Feature and other content.
Added the Search Feature, with both memoize (it probably doesn't need this but I wanted to test memoization in React - useMemo) and debouncing with 300ms.
I removed the case for N/A for the capitals I thought Nauru doesn't have a capital apparently it does now...
Restructured the app.jsx and added a new component Country Page that holds the sidebar, countrylist and searchbar. This was needed for easier prop passing to find the filtered countries from the search bar.